client: add subtokens
A token can be split into several subtokens, e.g. an address is composed
of a name, a street, ...
If a subtoken is created, its key is created from its partent's key and
its own subkey, e.g. if the parent has the key 'par' and the subtoken has
the subkey 'sub', the subtoken's key will be 'par-sub'